+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.1.6] - 2026-03-25
+### Changed
+- Update deprecated and obsolete APIs available in Unity 6000.4 and above.
+- Remove compatibility errors when using URP 2D renderer.
+
 ## [5.1.5] - 2026-02-12
 ### Changed
-- Migrate Mono APIs to CoreCLR-compatible APIs available in Unity 6000.5 and above
+- Migrate Mono APIs to CoreCLR-compatible APIs available in Unity 6000.5 and above.
 
 ## [5.1.4] - 2026-01-08
 ### Changed